agents/langraph-agent/requirements.txt
fastapi
uvicorn[standard]
pydantic
langgraph
langchain-google-genai
google-generativeai
python-multipart
httpx
python-dotenv